refactor: break down `rpc.rs`, extract ethereum and near into separate modules
